change: refactor filter to be more layout flexible and context based
