+++ .storybook/main.ts
@@ -1,11 +1,16 @@
 import type { StorybookConfig } from '@storybook/react-vite'
+
 const config: StorybookConfig = {
   stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: ['@storybook/addon-links', '@storybook/addon-docs'],
-
   framework: {
     name: '@storybook/react-vite',
     options: {}
+  },
+  docs: {},
+  typescript: {
+    reactDocgen: 'react-docgen'
   }
 }
+
 export default config
