+++ .storybook/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../tsconfig.app.json",
+  "compilerOptions": {
+    "allowJs": true,
+    "noEmit": true
+  },
+  "include": [
+    "**/*",
+    "../src/**/*"
+  ]
+}
