+++ package.json
@@ -86,5 +86,6 @@
       "npm run lint"
     ],
     "*.{css,md}": "npm run format"
-  }
+  },
+  "packageManager": "yarn@1.22.21+sha1.1959a18351b811cdeedbd484a8f86c3cc3bbaf72"
 }
