+++ src/components/_internal/base-checkbox/index.tsx
@@ -1,0 +1,29 @@
+import React, { type InputHTMLAttributes } from 'react'
+
+export type CheckboxProps = Pick<
+  InputHTMLAttributes<HTMLInputElement>,
+  | 'value'
+  | 'name'
+  | 'onChange'
+  | 'checked'
+  | 'defaultChecked'
+  | 'aria-checked'
+  | 'placeholder'
+  | 'type'
+>
+
+const BaseCheckbox: React.FC<CheckboxProps> = ({ placeholder, ...props }) => (
+  <div className="rsql-checkbox-wrapper">
+    <input
+      type="checkbox"
+      id="rsql-checkbox-id"
+      className="rsql-checkbox"
+      {...props}
+    />
+    <label htmlFor="rsql-checkbox-id" className="rsql-checkbox-label">
+      {placeholder}
+    </label>
+  </div>
+)
+
+export { BaseCheckbox }
