+++ src/index.ts
@@ -1,6 +1,4 @@
 import './styles.css'
-import { FiltersProvider } from './providers/FiltersProvider'
-export * from './components/Filter'
+export * from './components/filter'
 export * from './utils'
 export * from './types'
-export default FiltersProvider
