--- src/storybook/ComponentsWrapper.tsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import { ComponentsProvider } from '../providers/ComponentsProvider.tsx'
-
-const ComponentsWrapper = (Story: React.FC) => (
-  <ComponentsProvider>
-    <Story />
-  </ComponentsProvider>
-)
-
-export { ComponentsWrapper }
